+++ src/main/scala/com/testerhome/appcrawler/SuiteToClass.scala
@@ -12,14 +12,15 @@ object SuiteToClass extends CommonLog {
 
   var index=0
 
-  //todo: 特殊字符处理
+  //todo: 使用白名单
   def format(name:String): String ={
     name
       .replaceAllLiterally("\\", "\\\\")
       .replaceAllLiterally("\"", "\\\"")
       .replaceAllLiterally("#", "")
       .replaceAllLiterally("&", "")
-      .replaceAll("[#;/]", "")
+      .replaceAllLiterally("-", "")
+      .replaceAll("[#;/\\:]", "")
   }
   /**
     * 生成用例对应的class文件，用于调用scalatest执行
@@ -28,19 +29,25 @@ object SuiteToClass extends CommonLog {
   def genTestCaseClass(className: String, superClassName: String, fields: Map[String, Any], directory: String): Unit = {
     val pool = ClassPool.getDefault
     val classNameFormat = format(className)
+    log.trace(s"classNameFormat=${classNameFormat}")
     Try(pool.makeClass(classNameFormat)) match {
       case Success(classNew) => {
         classNew.setSuperclass(pool.get(superClassName))
         val init = new CtConstructor(null, classNew)
         val body = fields.map(field => {
+          //todo: 字段与类名分开
+          //todo: 使用数据驱动解决
           s"${field._1}_$$eq(${'"' + format(field._2.toString) + '"'}); "
         }).mkString("\n")
+        log.trace(body)
         init.setBody(s"{ ${body}\naddTestCase(); }")
         classNew.addConstructor(init)
         classNew.writeFile(directory)
       }
       case Failure(e) => {
         log.error(s"makeClass error with ${className}")
+        log.error(e.getMessage)
+        log.error(e.getStackTrace.mkString("\n"))
       }
     }
   }