+++ src/main/scala/com/testerhome/appcrawler/plugin/ReportPlugin.scala
@@ -35,11 +35,11 @@ class ReportPlugin extends Plugin with Report {
     val count=getCrawler().store.clickedElementsList.length
     log.info(s"clickedElementsList size = ${count}")
     val curSize=getCrawler().store.clickedElementsList.size
-    if(curSize-lastSize > curSize/10+20 ){
+/*    if(curSize-lastSize > curSize/10+20 ){
       log.info(s"${curSize}-${lastSize} > ${curSize}/10+20  ")
       log.info("generate test report ")
       generateReport()
-    }
+    }*/
   }
 
   def generateReport(): Unit ={